% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{importDemeter}
\alias{importDemeter}
\title{Import Demeter csv output}
\usage{
importDemeter(path)
}
\arguments{
\item{path}{The path to a Demeter csv file}
}
\value{
A data frame of the Demeter output
}
\description{
Read in the csv input for Demeter land allocation.
}
